requirements.txt
flask
apscheduler
sqlalchemy
easysnmp
pyyaml
pytz
-i https://mirrors.aliyun.com/pypi/simple/